requirements.txt
python-telegram-bot
pydub
speechrecognition
requests